bumps and fixes, oh my!
Bump the GDS version to the soon-to-be-released 1.7 version.

Finally fix the lingering memory leak...it's because I was reserving
memory in one of the allocators. Turns out that doesn't get freed
automatically, so either you just don't reserve any OR you free it
yourself.

Cleaned up and fixed some flushing issues...simplified it to a single
Flush task using a blocking deque. Meh. Might be faster? Might be
slower? Who cares, it's a LOT cleaner.