Fix race condition btw job status and flight info
Job status should be flipped to PRODUCING only after the flight info map
is updated, otherwise clients will freak out. Add an INITIALIZING status
as well to be a bit more granular.

More work is needed here...had to break the existing API for Jobs to let
the Producer flip the status. I don't like this but meh.