try compressed bitmaps to simplify streams
It seemed that previously the approach of simply filtering edges was
much faster than dealing with "node histories" etc. Since we're dealing
with a very sparse space (the set of all 64 bit values vs. the number of
edges for a given khop...so at most maybe 1e8 vs 1e18 or so) it might
make sense to use RoaringBitmap implementations.