Switch to a buffer passing model
Removes the need for semaphores and use a queue of available buffers
that, when "full", get added to the work queue for the flusher thread.
This allows the flusher to create the transfer pair and take ownership
of the Arrow bufs. The number of allocators stays fixed and we no longer
need to synchronize most memory events.

Still needs some thought on how to minimize blocking due to starvation,
but showing some promise. Plus it helped simplify some of the code :-)