fix a bug producing duplicate relationships
If submitting a GDS read job to export relationships and NOT specifying
at least a single relationship property (i.e. trying to use the behavior
of defaulting to ALL known properties) BUT there IS a property in the
projection, you'll get duplicate relationships returned for some
null/placeholder value.

This change changes the logic of assembling "triples" (triples is safe)
and prevents adding bogus triples to the array we use for generating the
stream.