wip redesign for supporting bulk import
Not complete yet as memory ownership isn't properly being passed to the
GDS graph but this work is needed as part of supporting bulk import.

The redesign puts responsibility of assembling batches into the WriteJob
implementation and not in the Producer. Sort of in the inverse of
producing streams, a WriteJob provides the Producer a Consumer for
offloading the batches of Vectors coming off the VectorSchemaRoot.