Fix memory leaks and closing GdsWriteJobs
This fixes up some lifecycle issues in the GdsWriteJob implementation
and should help clean up any memory and allocators. It still relies on
any GDS graphs to be dropped manually prior to shutdown, but now that
works again as before the redesign.

In addition, this now includes some state changes for GdsWriteJob
instances, which means clients can now wait for confirmation the job is
complete before moving to something like writing relationships.

In reality, the job should be 1 instance with 1 ticket and 2 calls to
put stream...but that's next.