start some tooling for testing via GDS csv export
In order to test database import, need someting to read import files.
This stubs out some basic PyArrow stuff to read a directory of GDS
import files that are similar to the ones used for bulk-import.

Array types are not supported due to PyArrow's CSV reader lacking
support for them.